refactor: ToggleButton component 이름 수정
- BackgroundTypeSelectButton으로 이름 수정
